requirements.txt
Django
djangorestframework
whitenoise
gunicorn
